+++ NEWS.md
@@ -1,5 +1,9 @@
 # ellmer (development version)
 
+* `chat_databricks()` now picks up on Databricks workspace URLs set in the
+  configuration file, which should improve compatibility with the Databricks CLI
+  (#521, @atheriel).
+
 # ellmer 0.2.0
 
 ## Breaking changes